requirements.txt
numpy
pillow
pywavelets
scipy
scikit-learn
tqdm
qrcode
reedsolo
pyzbar
